Update r.yml
I'll be honest, not sure why this works or if this is preferable. It's certainly less legible but is suggested by R's `usethis::use_github_action()`. It doens't seem as rooted in https://github.com/r-lib/actions/tree/v2/setup-r but it does work for some reason that is beyond me. So I reverted to this as I wasn't sure if v2 is up to date. I think this change is trying to pull the latest release of the r-lib actions above rather than hard-coding to v2 as I had here before.